+++ pdf/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.1
+
+- Fix Checkbox Widget
+
 ## 2.1.0
 
 - Fix SVG fit alignment